+++ GAM200/Engine/Engine/Code/ECS/Component/Graphics/CameraComponent.cpp
@@ -23,13 +23,13 @@ Technology is prohibited.
 namespace Engine
 {
     // Contructor for Camera Component
-    CameraComponent::CameraComponent(Entity_id _ID, int _height, float _fov, float _ar, bool _active) :
+    CameraComponent::CameraComponent(Entity_id _ID, float _height, float _fov, float _ar, bool _active) :
         IComponent{ _ID }, height{ _height }, fov{ _fov }, ar{ _ar }, isActive{ _active } {}
 
     // Deserialize function for Camera Component
     CameraComponent& CameraComponent::Deserialize(const DSerializer& _serializer)
     {
-        height = _serializer.GetValue<int>("Height");
+        height = _serializer.GetValue<float>("Height");
         fov = _serializer.GetValue<float>("FOV");
         ar = _serializer.GetValue<float>("AR");
