+++ GAM200/Engine/Engine/Header/ECS/System/CameraSystem.hpp
@@ -48,7 +48,7 @@ namespace Engine
 	private:
 		Math::vec2 camPosition{};
 
-		int camHeight{};
+		float camHeight{};
 		float camFov{};
 
 		GLFWwindow* pwindow{};
